Move broker interface holder to dataactcore.
To combat the integration tests issues we're having with
the broker vs validator interface holder, start the process
of combining them. Both are deprecated and will be removed
once the entire app is factored to use the GlobalDB object.
In the meantime, however, we need tests to work.
